Move buzzer to own folder
